repo sync: Handle many merged, branch delete: support multiple (#429)
Previously, `repo sync` would invoke `branch delete`
for each merged branch in an indeterminate order,
which would first move the upstacks of that branch to its base,
then delete the branch.

When multiple branches are merged, this has issues:

- repeated work of constantly moving the same branches
- log noise
- in some corner cases, (e.g. #398), the order of deletion matters
  and it causes the command to fail

The fix for this is multi-fold:

- `branch delete`: now accepts multiple branches
  instead of just one branch.
- `branch delete`: delete branches in reverse topological order.
  This will ensure that branches that have no dependents are deleted
  before branches that depend on them.
- `branch delete`: when multiple branches are being deleted,
  redirect the base of the upstacks to the next available branch,
  not just the immediate base.
- `repo sync`: use the multi-delete to delete merged branches

Resolves #398

+++ repo_sync.go
@@ -465,20 +465,26 @@ func (cmd *repoSyncCmd) deleteBranches(
 	remote string,
 	branchesToDelete []branchDeletion,
 ) error {
-	// TODO:
-	// Should the branches be deleted in any particular order?
-	// (e.g. from the bottom of the stack up)
-	for _, branch := range branchesToDelete {
-		err := (&branchDeleteCmd{
-			Branch: branch.BranchName,
-			Force:  true,
-		}).Run(ctx, log, opts)
-		if err != nil {
-			return fmt.Errorf("delete branch %v: %w", branch, err)
-		}
+	if len(branchesToDelete) == 0 {
+		return nil
+	}
 
-		// Also delete the remote tracking branch for this branch
-		// if it still exists.
+	branchNames := make([]string, len(branchesToDelete))
+	for i, b := range branchesToDelete {
+		branchNames[i] = b.BranchName
+	}
+
+	err := (&branchDeleteCmd{
+		Branches: branchNames,
+		Force:    true,
+	}).Run(ctx, log, opts)
+	if err != nil {
+		return fmt.Errorf("delete merged branches: %w", err)
+	}
+
+	// Also delete the remote tracking branch for this branch
+	// if it still exists.
+	for _, branch := range branchesToDelete {
 		remoteBranch := remote + "/" + branch.UpstreamName
 		if _, err := repo.PeelToCommit(ctx, remoteBranch); err == nil {
 			if err := repo.DeleteBranch(ctx, remoteBranch, git.BranchDeleteOptions{
